# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyRepo API",
        "description": "Documentación generada con l5-swagger",
        "contact": {
            "email": "soporte@midominio.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://myrepo.test",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/admin/categories": {
            "get": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Listado de categorías (paginado y filtrado).",
                "description": "Filtros:\n- business_id (recomendado)\n- parent_id (para listar subcategorías)\n- search (coincide en name o slug)\n- order_by: position|name|created_at (default: position)\n- sort: asc|desc (default: asc)\n- per_page, page",
                "operationId": "71fce704423c89eab31adce4a23df70b",
                "parameters": [
                    {
                        "name": "business_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "parent_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "order_by",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "position",
                                "name",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "data": [
                                        {
                                            "id": 10,
                                            "business_id": 1,
                                            "parent_id": null,
                                            "name": "Ropa",
                                            "slug": "ropa",
                                            "position": 1,
                                            "created_at": "2025-08-01T12:00:00Z",
                                            "updated_at": "2025-08-01T12:00:00Z",
                                            "children_count": 2,
                                            "products_count": 14,
                                            "parent": null,
                                            "business": {
                                                "id": 1,
                                                "name": "Mi Negocio"
                                            }
                                        }
                                    ],
                                    "meta": {
                                        "current_page": 1,
                                        "last_page": 1,
                                        "per_page": 20,
                                        "total": 1
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Crear una categoría.",
                "description": "Reglas:\n- `slug` único por (business_id, slug). Si no se envía, se genera desde `name`.\n- `parent_id` (si se envía) debe pertenecer al mismo `business_id`.\n- Se valida **límite de categorías** según la suscripción activa del negocio.",
                "operationId": "810e8b9fee1619d4229208cb942c486a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "business_id",
                                    "name"
                                ],
                                "properties": {
                                    "business_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Ropa"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "ropa",
                                        "nullable": true
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 10,
                                    "business_id": 1,
                                    "parent_id": null,
                                    "name": "Ropa",
                                    "slug": "ropa",
                                    "position": 1
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación / Límite alcanzado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "message": "Has alcanzado el límite de categorías para tu plan."
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/{category}": {
            "get": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Mostrar una categoría por ID.",
                "operationId": "11a8aaa127a831f1b2178b277adfce70",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 10,
                                    "business_id": 1,
                                    "parent_id": null,
                                    "name": "Ropa",
                                    "slug": "ropa",
                                    "position": 1,
                                    "created_at": "2025-08-01T12:00:00Z",
                                    "updated_at": "2025-08-01T12:00:00Z",
                                    "children_count": 2,
                                    "products_count": 14,
                                    "parent": null,
                                    "business": {
                                        "id": 1,
                                        "name": "Mi Negocio"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Eliminar una categoría.",
                "description": "Efectos por FKs:\n- `parent_id` de los hijos se pone en NULL (SET NULL)\n- Relaciones en `product_category` se eliminan (CASCADE)",
                "operationId": "1c50d201211dea34943f752a3cbb8e87",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Admin - Categories"
                ],
                "summary": "Actualizar una categoría.",
                "description": "Notas:\n- No se permite cambiar `business_id`.\n- `parent_id`, si se envía, debe ser del mismo negocio y no puede ser el propio id\nni un descendiente (ciclo). Se valida de forma simple subiendo por ancestros.",
                "operationId": "8fd4e8d94872fea0b6efac980d6f6a6f",
                "parameters": [
                    {
                        "name": "category",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "parent_id": {
                                        "type": "integer",
                                        "example": null,
                                        "nullable": true
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Ropa y Accesorios"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "ropa-accesorios"
                                    },
                                    "position": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 10,
                                    "business_id": 1,
                                    "parent_id": null,
                                    "name": "Ropa y Accesorios",
                                    "slug": "ropa-accesorios",
                                    "position": 2
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "403": {
                        "description": "No autorizado"
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Bearer",
                "scheme": "bearer"
            }
        }
    },
...
}